Refactor into hook
